refact: use Robolectric at :composite:timeline-test
