+++ composite/timeline-test/src/test/java/br/com/orcinus/orca/composite/timeline/test/post/time/StringRelativeTimeProviderTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023-2024 Orcinus
+ * Copyright © 2023–2024 Orcinus
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
