+++ composite/timeline-test/src/test/java/br/com/orcinus/orca/composite/timeline/test/refresh/SemanticsMatcherExtensionsTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023-2024 Orcinus
+ * Copyright © 2023–2024 Orcinus
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -13,21 +13,24 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.composite.timeline.test.stat.favorite
+package br.com.orcinus.orca.composite.timeline.test.refresh
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
-import br.com.orcinus.orca.composite.timeline.stat.Stats
-import br.com.orcinus.orca.composite.timeline.test.stat.activateable.favorite.onFavoriteStat
-import br.com.orcinus.orca.platform.autos.theme.AutosTheme
+import br.com.orcinus.orca.composite.timeline.Timeline
+import br.com.orcinus.orca.composite.timeline.refresh.Refresh
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 
-internal class SemanticsNodeInteractionsProviderExtensionsTests {
+@RunWith(RobolectricTestRunner::class)
+internal class SemanticsMatcherExtensionsTests {
   @get:Rule val composeRule = createComposeRule()
 
   @Test
-  fun findsFavoriteStat() {
-    composeRule.apply { setContent { AutosTheme { Stats() } } }.onFavoriteStat().assertIsDisplayed()
+  fun matchesNodeThatIsInProgress() {
+    composeRule.setContent { Timeline(onNext = {}, refresh = Refresh.Indefinite) {} }
+    composeRule.onNode(isInProgress()).assertIsDisplayed()
   }
 }
